chore(app): more modular folders architecture
