+++ src/store/getters.js
@@ -1,0 +1,1 @@
+// export global getters out there ?
